Update points entry to allow undefined
The round controller should accept undefined from NumberEntry to
indicate the user hasn't played a tile. This is needed because there is
tile with a value of 0 and it is needed to distinguish a played tile
with a value 0 from a turn where a player hasn't played any tiles
at all.