Make FlyIn responsible for the hidden attribute
The FlyIn controller should be responsible for manipulating the hidden
attribute of the base element. If this is not done by FlyIn each
consumer of the class needs to add the timeout construction to ensure
the classes aren't manipulated before the browser has had a chance to
process the new hidden value.