Fix round won detection
The round controller was checking whether or not the current player has
an empty hand. This check was no longer valid since updating the score
has been removed from the round controller. The tile delta hasn't been
processed yet at the point where round controller is checking for an
empty hand.

The empty hand check has been moved to the main module. This is where
the score is processed and thus the first place where a valid check for
an empty hand can be performed.

Due to the round controller no longer checking for an empty hand the
RoundStatus won is no longer needed.